src/main/java/com/shabloel/recipeapp/services/RecipeServiceImpl.java
package com.shabloel.recipeapp.services;


import com.shabloel.recipeapp.exception.RecipeNotFoundException;
import com.shabloel.recipeapp.model.Recipe;
import com.shabloel.recipeapp.repositories.RecipeRepository;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.web.bind.annotation.RequestParam;

import java.util.HashSet;
import java.util.Set;

/**
 * @author : christiaan.griffioen
 * @since :  31-3-2021, wo
 **/
@Slf4j
@Service
public class RecipeServiceImpl implements RecipeService {

    private final RecipeRepository recipeRepository;

    @Autowired
    public RecipeServiceImpl(RecipeRepository recipeRepository) {
        this.recipeRepository = recipeRepository;
    }

    @Override
    public Set<Recipe> getRecipes() {
        Set<Recipe> recipeSet = new HashSet<>();
        recipeRepository.findAll().iterator().forEachRemaining(recipeSet::add);
        return recipeSet;
    }

    @Override
    public Recipe getRecipeById(Long id) {
        return recipeRepository.getById(id);
    }

    @Override
    public void addNewRecipe(Recipe recipe) {
        recipeRepository.save(recipe);
    }

    @Override
    public void deleteRecipe(Long id) {
        if (!recipeRepository.existsById(id)) {
            throw new RecipeNotFoundException("Recipe with id " + id + " does not exists");
        }
        recipeRepository.deleteById(id);
    }
}
